Made the cancel button optional

+++ src/main/java/org/cascadebot/cascadebot/utils/ConfirmUtils.java
@@ -34,7 +34,7 @@ public class ConfirmUtils {
     private static ListMultimap<String, ConfirmRunnable> confirmedMap = ArrayListMultimap.create();
 
     //region Confirm Action
-    public static boolean confirmAction(long userId, String actionKey, TextChannel channel, MessageType type, String message, long buttonDelay, long expiry, ConfirmRunnable action) {
+    public static boolean confirmAction(long userId, String actionKey, TextChannel channel, MessageType type, String message, long buttonDelay, long expiry, boolean isCancellable, ConfirmRunnable action) {
         GuildData guildData = GuildDataManager.getGuildData(channel.getGuild().getIdLong());
         boolean useEmbed = guildData.getCoreSettings().isUseEmbedForMessages();
         Message sentMessage;
@@ -60,10 +60,12 @@ public class ConfirmUtils {
                     if (runner.getIdLong() != action.userId) return;
                     action.run();
                 }));
-                group.addButton(new Button.UnicodeButton(UnicodeConstants.RED_CROSS, ((runner, channel1, message1) -> {
-                    confirmedMap.remove(actionKey, action);
-                    sentMessage.delete().queue(null, DiscordUtils.handleExpectedErrors(ErrorResponse.UNKNOWN_MESSAGE));
-                })));
+                if (isCancellable) {
+                    group.addButton(new Button.UnicodeButton(UnicodeConstants.RED_CROSS, ((runner, channel1, message1) -> {
+                        confirmedMap.remove(actionKey, action);
+                        sentMessage.delete().queue(null, DiscordUtils.handleExpectedErrors(ErrorResponse.UNKNOWN_MESSAGE));
+                    })));
+                }
                 group.addButtonsToMessage(sentMessage);
                 group.setMessage(sentMessage.getIdLong());
                 guildData.addButtonGroup(channel, sentMessage, group);
@@ -78,7 +80,7 @@ public class ConfirmUtils {
         return true;
     }
 
-    public static void confirmAction(long userId, String actionKey, TextChannel channel, MessageType type, String message, ConfirmRunnable action) {
+    public static void confirmAction(long userId, String actionKey, TextChannel channel, MessageType type, String message, boolean isCancellable, ConfirmRunnable action) {
         confirmAction(
                 userId,
                 actionKey,
@@ -87,11 +89,12 @@ public class ConfirmUtils {
                 message,
                 TimeUnit.SECONDS.toMillis(2),
                 TimeUnit.MINUTES.toMillis(1),
+                isCancellable,
                 action
         );
     }
 
-    public static void confirmAction(long userId, String actionKey, TextChannel channel, String message, ConfirmRunnable action) {
+    public static void confirmAction(long userId, String actionKey, TextChannel channel, String message, boolean isCancellable, ConfirmRunnable action) {
         confirmAction(
                 userId,
                 actionKey,
@@ -100,6 +103,7 @@ public class ConfirmUtils {
                 message,
                 TimeUnit.SECONDS.toMillis(2),
                 TimeUnit.MINUTES.toMillis(1),
+                isCancellable,
                 action
         );
     }
